requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
Pillow==11.3.0

# MLOps Optimization: Force CPU-only PyTorch to shrink Docker image size
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.1+cpu
torchvision==0.16.1+cpu
numpy==1.26.4
opencv-python-headless==4.11.0.86